Remove bundled PGoApi package and replace with dependency.
Remove the entire PGoApi package from being distributed through
source control. Replace with single listed dependency in the
requirement.txt file. Additionally, add a setup.py file for easier
installation and setup of this package, and pull dependencies from
the now updated requirement.txt file.

As this fork is a proposal for modifications to the upstream
repository, I used the original title, description, and author
from the original repo to populate this package metadata.