Restructure project into main package and visualization templates.
Move all scripts into a new package, and all html files into a
templates directory. Propose in the future to either come up with
some way to populate the API keys from the config file, or if
necessary the html files could be replaced with something like
Jinja2 templates and the API keys populated at runtime.

THIS COMMIT IS NOT FUNCTIONAL. Will add new functionality to
call the resturctured code from the command line.