Fix requirement.txt (again).
It was a mistake to add the -e flag. The PGoApi package is now
properly installed. Add an editable install line of the current
directory package to requirements.txt so that the spawnscan
pacakge in the top level directory is installed, installing
further dependencies and creating console script entry point.

NOTE:  As of this point, while not fully tested, it seems to
       now be Python3 compatible.