Add multiprocessing hex scan.
Implemented a mutiprocessing scan using a Pool, which reduces the
scan time of an approximatly 5km^2 area from about 50 seconds to
6 or 7 (the 8-fold improvement makes sense as testing is being done
on an 8 core machine; improvements will likely only scale with the
number of processor cores).

Updated the console module to call the new multiprocessor scan, and
added a bit of whitespace and comments to the check module.