Update multiscan to use each user as a process.
The multiprocessing Pool spawns a worker process for each user,
which now seems fine as having multiple process for a single user
now seems unnecessary as queries per user are being throttled on
the server side. Update check algorithm to take into account the
throttling.

Implement a generator wrapping a Queue to feed the pool imap call.
This alows for failed scans to be feed back into the queue to be
tried again by another worker (or the same) at a future time.
Upon success, a synchronized counter is incremented, and the queue
wrapper then can just keep feeding the pool scan positions until
the number of successes is equal to the input number (all failures
where re-scanned successfully). There did seem to be a little race
condition at some point that would hang the program, but I think
it had to do with the Lock for the counter value not being
implemented properly, as when I added the same lock to each scanner
function and used that so synchronize the success counter, the
problem seemed to go away (meaning it went from happening nearly
every time to not yet repeating in the numerous tests since fixing.)
So I can't prove that it's working entirely properly, but
experimental evidence suggests my issue was corrected.